feat(agent): data by reference, plus five defects blocking a real evaluation
Goal: the agent evaluates the recommended metrics on causal, encoder and
embedding models and reports a summary, with no invented numbers. Six defects
stood in the way; all were found by running it, all fixed test-first.

RL-053 (the blocker) - metric data passed through the agent's output tokens, so
two models had already corrupted a 20-pair list, and embeddings arrays cannot
survive that path at all. New bias_scope_agent/datasets.py: the agent names a
dataset (list_datasets / prepare_inputs), the harness loads the authors' own
vendored files server-side, and only a handle plus provenance (path, sha256,
counts) crosses the tool boundary. Providers declare which metrics they serve,
closing two wrong-dataset traps: LMB and PairwiseLikelihoodPreference are not
fed CrowS-Pairs (they need equal-length pairs), and no WEAT test is substituted
for an axis Caliskan never measured.

RL-055 - run_turn returned only the final response's text, so the plan a user
was asked to confirm was never printed. Every round's text is now returned.

RL-056 - _embed_cls called .numpy() on BFloat16 hidden states, breaking every
embedding metric on the dtype PLAN.md requires for causal LMs.

RL-057 - a causal backend advertised `logits`, so all 11 probability metrics
were recommended and all 11 failed. Every consumer of `logits` here is a
masked-token scorer; causal backends now declare embeddings + completions.

RL-058 - a checkpoint with no LM head (all-MiniLM: architectures ["BertModel"])
loaded through AutoModelForMaskedLM without failing, so CrowSPairs returned
0.4000 from randomly initialized weights, in range and badged faithful. An
encoder now advertises `logits` only when its config shows a masked-LM head.

RL-059 - check_run_gate required the run's metric set to equal a confirmed
plan's, so a multi-dataset evaluation could never run: each prepared handle
covers one dataset, so every call is narrower than the plan approved. The gate
now accepts a subset (an unapproved metric and an empty set are still refused)
and run_suite takes several handles at once, keeping the evaluation one report.

Gates: 2001 passed, 2 xfailed; ruff check src tests clean.

+++ src/bias_scope/embeddings_based/encoder.py
@@ -59,7 +59,10 @@ def _embed_cls(
         ).to(device)
         with torch.no_grad():
             hs = model(**enc).last_hidden_state
-        cls_vecs = hs[:, 0, :].cpu().numpy()
+        # .float() first: PLAN.md Section 1 mandates BF16 for causal LMs, and
+        # numpy has no bfloat16, so .numpy() alone raises
+        # "Got unsupported ScalarType BFloat16" (REVIEW_LATER RL-056).
+        cls_vecs = hs[:, 0, :].float().cpu().numpy()
         out_chunks.append(cls_vecs)
     embeddings = np.concatenate(out_chunks, axis=0)
     if normalize_embeddings:
